Completed the basic component model
Written the basic component model system. There is a base Mud class that acts as a container for Components and Services. Components are created and passed the Mud they belong to, where they register themselves with the mud and any services they, or sub components, provide.

When the mud is asked to Run via Mud.Run(), the components are all initialized. This is where they request services exposed by other components.