Logging component written
Basic component that setup up a LogDispatcher that initially dispatches log data to file.

The log dispatcher is both a ILogDispatcher and ILogger, but must contain other ILoggers in order to log anything. Other ILoggers can be added and removed from the LogDispatcher and this can be use to add other Log destinations at runtime.