Initial Command base class and CommandArguments class written
CommandArguments parses a command line string it's given into it's components. Using the standard split on whitespace unless contained in quotes.

However the parsing is done in the constructor right now. I don't like it so I'm going to move it into a factory static method I think.

Command is a simple base class, with an Execute() method, derived commands will have to specify a type for the class using generics. This is to set what that command acts on / with. For example some commands may work with the Session object instead of a Player object. This will be reconciled when I get down to creating the CommandManager class and CommandContext classes.

Command's are also constructed with IServiceProvider, this will be because most commands will probably need to access mud services in order to perform their function.