Command system in place.
Commands inherit from Command<T> which in turn implements ICommand.
Commands also have a CommandAttribute which the name of the command is specified.

The T generic argument in Command<T> donate the type that is the "Actor" of the command.

To register a command with the system, retrieve the service ICommandRegistrar and call ICommandRegistrar.Register passing in a System.Type of the command to register.

To call commands, first you need to create a CommandContext via the ICommandContextFactory service, ideally you want a single command context per user / command usage context. Then you need to register one or more "Actors" with the context.

Then you need to create a CommandArguments by calling CommandArguments.Create(string commandLine).

Then pass the create CommandArguments into CommandContext.FindCommand.

If an ICommand is returned, the command was found and you can just call ICommand.Execute(), if null was returned then no command was found and you should probably let the user know.