Fixed a bug in the session input reader, where if the connection was changed it was falling over when it tried to restart itself.
This was due to the fact that if the ConnectionChanged event is called on the same thread as the input thread, then stop doesn't actually stop the thread and thus when it called Start() again it failed with InvalidOperationException stating the input reader was already started.

Added code then, to have a different way of changing the currently watched connection if you happened to enter on the same thread.