Add the Session.TakeControlOf(Session) method
However as the comment in the code states, it's badly named. Need a better name for it.

It basically swaps the current sessions IConnection with the IConnection of the passed in session, then closes the old connection and the passed in session. Effectively making it seem to the user that the passed in session has taken control of the current session.