Added AdminLoginManager
The AdminLoginManger provides a service that allows you to perform the login procedure for a given admin user and session.

It keeps track of currently logged in users so that active session retrieval is possible.

LoginState has been changed to use the AdminLoginManager and AdminComponent registers it as a service.